A Simple Password generator
In this program, the random library is imported. A few sets of characters are entered for the system to choose from. The user is asked for the inputs like the password length, numbers of special characters, etc. once the input is given, the system selects random items from the list and displays the final password.